Complete empolyee class
The task is to extend employee.py such that all the tests pass. The solution is able to calculate total pay from the parameters provided and produce a string that explains how pay is calculated. The solution constructs appropriate employee objects for the six employees identified in the employee.py file.